move/add: profile edit 컴포넌트 분리(profile image 분리)
