+++ src/types/supabase.ts
@@ -6,7 +6,6 @@ export type Json =
   | { [key: string]: Json | undefined }
   | Json[]
 
-
 export type Database = {
   public: {
     Tables: {
@@ -219,7 +218,12 @@ export type Database = {
       [_ in never]: never
     }
     Functions: {
-      [_ in never]: never
+      delete_user: {
+        Args: {
+          user_id: string
+        }
+        Returns: undefined
+      }
     }
     Enums: {
       [_ in never]: never
@@ -230,7 +234,6 @@ export type Database = {
   }
 }
 
-export type InfomationTable = Database['public']['Tables']['information']['Insert']
 type PublicSchema = Database[Extract<keyof Database, "public">]
 
 export type Tables<
@@ -312,4 +315,3 @@ export type Enums<
   : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
     ? PublicSchema["Enums"][PublicEnumNameOrOptions]
     : never
-
